+++ src/Point.cpp
@@ -5,7 +5,7 @@ Point::Point()
     //ctor
 }
 
-Point::Point(float x, float y)
+Point::Point(double x, double y)
 {
     this->x = x;
     this->y = y;
@@ -17,26 +17,25 @@ Point::~Point()
 }
 
 
-float Point::getX()
+double Point::getX()
 {
     return this->x;
 }
 
-float Point::getY()
+double Point::getY()
 {
     return this->y;
 }
 
-void Point::setX(float x)
+void Point::setX(double x)
 {
     this->x = x;
 }
 
-void Point::setY(float y)
+void Point::setY(double y)
 {
     this->y = y;
 }
-
 
 double toRadians(double degrees)
 {
